reconcile: 回放本地未提交的 RFC-002 延续工作(gallery/音频/视频/前端/文档)

+++ hevi/assembly/subtitle_align.py
@@ -31,7 +31,7 @@ def _fmt_ts(seconds: float) -> str:
     """秒 → SRT 时间码 HH:MM:SS,mmm。"""
     if seconds < 0:
         seconds = 0.0
-    ms = int(round(seconds * 1000))
+    ms = round(seconds * 1000)
     h, ms = divmod(ms, 3_600_000)
     m, ms = divmod(ms, 60_000)
     s, ms = divmod(ms, 1000)
